migrate offset tests
